Initial setup of Node & Babel
