+++ .travis.yml
@@ -1,16 +1,24 @@
-language: python
+dist: trusty
 sudo: true
-python:
-  - "3.6"
+language: node_js
+node_js: 11
+
+cache:
+    - yarn
+    - pip
+    - directories:
+        - /opt/python/
+        - /home/travis/.local/share/
+
 # command to install dependencies
 install:
-    - pip install pipenv
+    - pip install --user -U pip pipenv
     - pipenv install --dev
 
 script:
+  - yarn compile
   - make test
 
 branches:
   only:
     - master
-    - frontend-bootstrap-feature-branch
