+++ README.md
@@ -2,6 +2,21 @@
 [![Build Status](https://api.travis-ci.org/ONSdigital/sdc-responses-dashboard.svg?branch=master)](https://travis-ci.org/ONSdigital/sdc-responses-dashboard)
 [![Codacy Badge](https://api.codacy.com/project/badge/Grade/80ad95f7aaa9477da6aa8fd9aec40f52)](https://www.codacy.com/project/MebinAbraham/sdc-responses-dashboard/dashboard?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=ONSdigital/sdc-responses-dashboard&amp;utm_campaign=Badge_Grade_Dashboard)
 
+### Front-end Toolkit
+
+The front-end toolkit uses nodejs, yarn and gulp.
+
+Currently, in order to build the front-end toolkit, you will need to have node. To do this, do the following commands:
+```
+brew install node
+```
+
+Install yarn with:
+
+```
+npm install yarn --global
+```
+
 ## Run the application
 
 **Initial Setup:**
@@ -15,6 +30,11 @@ pip install --U pip setuptools pipenv
 
 **Run with Make:**
 
+Compile static asssets with
+```bash
+make compile
+```
+
 Install dependencies with
 ```bash
 make build
@@ -26,6 +46,11 @@ make start
 ```
 
 **Alternatively, run manually:**
+
+Compile static assets:
+```bash
+yarn compile
+```
 
 Install dependencies to run:
 ```bash
