+++ app/templates/errors/error.html
@@ -8,7 +8,7 @@
 		<meta content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" name="viewport">
 		{% include 'partials/_styles.html' %}
 		<!-- Custom CSS -->
-		<link rel="stylesheet" href="{{ url_for('static', filename='assets/css/custom-errors.css') }}">
+		<link rel="stylesheet" href="{{ url_for('static', filename='dist/css/custom-errors.min.css') }}">
 	</head>
 	<body class="no-sidebar skin-black">
 		<!-- sidebar-collapse -->
