+++ app/templates/partials/_footer.html
@@ -2,7 +2,8 @@
     <div class="pull-right hidden-xs">
         <b>SDC Responses Dashboard</b>
     </div>
-    <strong><span id="time-updated-label" style="display: none;">Last Updated:</span>
-        <em><span id="time-updated">&nbsp;</span></em></strong>
+    <strong>
+        <span id="time-updated-label" style="display: none;">Last Updated:</span>
+        <em><span id="time-updated">&nbsp;</span></em>
     </strong>
 </footer>
