+++ app/templates/partials/_modal_collex.html
@@ -8,31 +8,26 @@
                 <h4 class="modal-title">Choose A Collection Exercise For <strong><span id="chosen-survey"></span></strong></h4>
             </div>
             <div class="modal-body">
-                <p>
-                    <!-- search form -->
-                    <div class="sidebar-form">
-                        <div class="input-group">
-                            <input type="text" id="collex-search" name="q" class="form-control" placeholder="Search...">
-                            <span class="input-group-btn">
-								<button type="submit" name="search" id="search-btn" class="btn btn-flat">
-								<i class="fa fa-search"></i>
-								</button>
-							</span>
+                <!-- search form -->
+                <div class="sidebar-form">
+                    <div class="input-group">
+                        <input type="text" id="collex-search" name="q" class="form-control" placeholder="Search...">
+                        <span class="input-group-btn">
+                            <button type="submit" name="search" id="search-btn" class="btn btn-flat">
+                            <i class="fa fa-search"></i>
+                            </button>
+                        </span>
+                    </div>
+                </div>
+                <!-- /.search form -->
+                <div id="collex_modal_wrapper" class="dataTables_wrapper form-inline dt-bootstrap">
+                    <div class="row">
+                        <div class="col-sm-12">
+                            <table id="collex-datatable" class="table table-bordered table-striped dataTable table-cell-border table-datatable-hover" role="grid" aria-describedby="surveys" style="clear:both;">
+                            </table>
                         </div>
                     </div>
-                    <!-- /.search form -->
-                    <div id="collex_modal_wrapper" class="dataTables_wrapper form-inline dt-bootstrap">
-                        <div class="row">
-                            <div class="col-sm-12">
-                                <table id="collex-datatable" class="table table-bordered table-striped dataTable table-cell-border table-datatable-hover" role="grid" aria-describedby="surveys" style="clear:both; height: auto !important;">
-                                <col style="width: 50%;" />
-                                <col style="width: 25%;" />
-                                <col style="width: 25%;" />
-                                </table>
-                            </div>
-                        </div>
-                    </div>
-                </p>
+                </div>
             </div>
             <div class="modal-footer">
                 <button type="button" class="btn btn-primary pull-left" data-dismiss="modal" data-toggle="modal" data-target="#modal-survey">Back</button>
