+++ app/templates/partials/_modal_survey.html
@@ -8,46 +8,44 @@
                 <h4 class="modal-title">Choose A Survey</h4>
             </div>
             <div class="modal-body">
-                <p>
-                    <!-- search form -->
-                    <div class="sidebar-form">
-                        <div class="input-group">
-                            <input type="text" id="survey-search" name="q" class="form-control" placeholder="Search...">
-                            <span class="input-group-btn">
-						<button type="submit" name="search" id="search-btn" class="btn btn-flat">
-						<i class="fa fa-search"></i>
-						</button>
-						</span>
+                <!-- search form -->
+                <div class="sidebar-form">
+                    <div class="input-group">
+                        <input type="text" id="survey-search" name="q" class="form-control" placeholder="Search...">
+                        <span class="input-group-btn">
+                    <button type="submit" name="search" id="search-btn" class="btn btn-flat">
+                    <i class="fa fa-search"></i>
+                    </button>
+                    </span>
+                    </div>
+                </div>
+                <!-- /.search form -->
+                <div id="survey_modal_wrapper" class="dataTables_wrapper form-inline dt-bootstrap">
+                    <div class="row">
+                        <div class="col-sm-12">
+                            <table id="survey-datatable" class="table table-bordered table-striped dataTable table-cell-border table-datatable-hover" role="grid" aria-describedby="surveys" style="clear:both;">
+                                <col style="width: 70%" />
+                                <col style="width: 30%" />
+                                <thead>
+                                    <tr role="row">
+                                        <th class="sorting_asc" tabindex="0" aria-controls="survey_modal" rowspan="1" colspan="1" aria-sort="ascending" aria-label="Survey Name" style="width: 70%;">Survey</th>
+                                        <th class="sorting" tabindex="0" aria-controls="survey_modal" rowspan="1" colspan="1" aria-label="Survey Reference No." style="width: 300px;">Reference</th>
+                                    </tr>
+                                </thead>
+                                <tbody>
+                                    {% for survey in all_surveys %}
+                                        {% if survey.collectionExercises %}
+                                            <tr role="row" class="even" id="{{ survey.surveyId }}" data-survey-short-name="{{ survey.shortName }}" data-survey-long-name="{{ survey.longName }}">
+                                                <td class="sorting_1">{{ survey.shortName }}<span id="long-name"> - {{ survey.longName }}</span></td>
+                                                <td>{{ survey.surveyRef }}</td>
+                                            </tr>
+                                        {% endif %}
+                                    {% endfor %}
+                                </tbody>
+                            </table>
                         </div>
                     </div>
-                    <!-- /.search form -->
-                    <div id="survey_modal_wrapper" class="dataTables_wrapper form-inline dt-bootstrap">
-                        <div class="row">
-                            <div class="col-sm-12">
-                                <table id="survey-datatable" class="table table-bordered table-striped dataTable table-cell-border table-datatable-hover" role="grid" aria-describedby="surveys" style="clear:both;">
-                                    <col style="width: 70%" />
-                                    <col style="width: 30%" />
-                                    <thead>
-                                        <tr role="row">
-                                            <th class="sorting_asc" tabindex="0" aria-controls="survey_modal" rowspan="1" colspan="1" aria-sort="ascending" aria-label="Survey Name" style="width: 70%;">Survey</th>
-                                            <th class="sorting" tabindex="0" aria-controls="survey_modal" rowspan="1" colspan="1" aria-label="Survey Reference No." style="width: 300px;">Reference</th>
-                                        </tr>
-                                    </thead>
-                                    <tbody>
-                                        {% for survey in all_surveys %}
-											{% if survey.collectionExercises %}
-												<tr role="row" class="even" id="{{ survey.surveyId }}" data-survey-short-name="{{ survey.shortName }}" data-survey-long-name="{{ survey.longName }}">
-													<td class="sorting_1">{{ survey.shortName }}<span id="long-name"> - {{ survey.longName }}</span></td>
-													<td>{{ survey.surveyRef }}</td>
-												</tr>
-											{% endif %}
-										{% endfor %}
-                                    </tbody>
-                                </table>
-                            </div>
-                        </div>
-                    </div>
-                </p>
+                </div>
             </div>
             <div class="modal-footer modal-survey-footer">
                 <button type="button" class="btn btn-primary pull-left" data-dismiss="modal">Close</button>
