+++ app/templates/partials/_plugins.html
@@ -1,0 +1,8 @@
+<!-- Moment JS -->
+<script src="//cdnjs.cloudflare.com/ajax/libs/moment.js/2.8.4/moment.min.js"></script>
+<!-- DataTables -->
+<script src="https://cdn.datatables.net/1.10.19/js/jquery.dataTables.min.js"></script>
+<script src="https://cdn.datatables.net/1.10.19/js/dataTables.bootstrap.min.js"></script>
+<script src="https://cdn.datatables.net/scroller/1.5.1/js/dataTables.scroller.min.js"></script>
+<!-- App -->
+<script src="{{ url_for('static', filename='dist/js/datatables.min.js') }}"></script>
