+++ app/templates/partials/_styles.html
@@ -5,8 +5,8 @@
 <link href="https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet" integrity="sha384-wvfXpqpZZVQGK6TAh5PVlGOfQNHSoD2xbE+QkPxCAFlNEevoEH3Sl0sibVcOQVnN"
 	crossorigin="anonymous">
 <!-- Theme style -->
-<link rel="stylesheet" href="{{ url_for('static', filename='assets/css/AdminLTE.min.css') }}">
-<link rel="stylesheet" href="{{ url_for('static', filename='assets/css/skins/skin-black.css') }}">
+<link rel="stylesheet" href="{{ url_for('static', filename='dist/css/AdminLTE.min.css') }}">
+<link rel="stylesheet" href="{{ url_for('static', filename='dist/css/skin-black.min.css') }}">
 <!-- Google Font -->
 <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,600,700,300italic,400italic,600italic">
 <!-- Favicon -->
