+++ app/templates/reporting.html
@@ -10,8 +10,8 @@
 		<!-- DataTables -->
 		<link rel="stylesheet" href="https://cdn.datatables.net/1.10.19/css/dataTables.bootstrap.min.css" />
 		<!-- Custom CSS -->
-		<link rel="stylesheet" href="{{ url_for('static', filename='assets/css/custom-reporting.css') }}" />
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/pace/1.0.2/pace.js"></script>
+		<link rel="stylesheet" href="{{ url_for('static', filename='dist/css/custom-reporting.min.css') }}" />
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/pace/1.0.2/pace.min.js"></script>
         <link href="https://cdnjs.cloudflare.com/ajax/libs/pace/1.0.2/themes/blue/pace-theme-flash.min.css" rel="stylesheet" />
 	</head>
 	<body id="collex-id" data-collex="{{ collex_id }}" data-collection-instrument-type="{{ collection_instrument_type }}" data-reporting-refresh-cycle="{{ reporting_refresh_cycle }}" class="no-sidebar skin-black">
@@ -96,15 +96,9 @@
 		</div>
 		<!-- ./wrapper -->
 		{% include 'partials/_scripts.html' %}
-		<!-- Moment JS -->
-		<script src="//cdnjs.cloudflare.com/ajax/libs/moment.js/2.8.4/moment.min.js"></script>
-		<!-- DataTables -->
-		<script src="https://cdn.datatables.net/1.10.19/js/jquery.dataTables.min.js"></script>
-		<script src="https://cdn.datatables.net/1.10.19/js/dataTables.bootstrap.min.js"></script>
-		<script src="https://cdn.datatables.net/scroller/1.5.1/js/dataTables.scroller.min.js"></script>
+		{% include 'partials/_plugins.html' %}
 		<!-- App -->
-		<script src="{{ url_for('static', filename='assets/js/dashboard.js') }}"></script>
-		<script src="{{ url_for('static', filename='assets/js/datatables.js') }}"></script>
+		<script src="{{ url_for('static', filename='dist/js/reporting.min.js') }}"></script>
 		<script>
 			$('#collex-id').data('surveys', '{{ all_surveys|tojson|safe }}');
 		</script>
